requirements.txt
openai
python_dotenv
discord